refactor(dhcp): extract DHCPv6 DUID management to duid.go
Continue the #6430 split by moving DHCPv6 DUID identity management out
of dhcp.go into a new sibling duid.go: the DUIDInfo display type, the
DUIDs() / ClearDUID() / ClearAllDUIDs() operations, getDUID() (load /
generate / persist with the #5855 type-rotation fail-safe), the
duidPath() / validInterfaceName() #4857 traversal guards, loadDUID() /
saveDUID() / the duidWriteFile persist seam, and the actualDUIDType() /
normalizeDUIDType() helpers.

Pure code-motion: whole functions, types, and the package-level
duidWriteFile var move verbatim with unchanged signatures and bodies;
only the per-file package clause and (goimports-pruned) import block
differ. duid.go is the natural home for the encoding/hex, os,
path/filepath, iana, and fsatomic dependencies, which now leave
dhcp.go's import set.

Validation: go build ./pkg/dhcp/, go vet ./pkg/dhcp/ clean.
